Add support to set visitor on request without a visit
Split the domain model of a request from the underlying database
structure.
In the domain a request contains a visitor. A calendar event can be
attached to a request representing the scheduled event in an Outlook
agenda.

The request domain model is spread over 2 classes, request and visit, in
the data provider layer. Creation and deletion of visits is handled by
the request data provider. The visit data provider only provides an
update endpoint through which details of the calendar event can be updated.