Add endpoint to restore VAT certificates in bulk
In Access VAT certificates were recycled using a pointer to the
original certificate in the DB. RKB expects the PDF file to be
copied.
An endpoint has been added to run in the background and make sure all
VAT certificates are copied for those only containing a pointer in the
DB at the moment.